#273: Move apply_colony_command to colony/remote.rs
Mechanical refactor flagged by #270 design review (Design #3). The ~250-
line arrival handler was sitting in `communication/mod.rs` but dealt
entirely with `colony/`-owned types (`BuildingOrder`, `DemolitionOrder`,
`UpgradeOrder`, `BuildOrder`), cost/time modifier application, and the
`BuildingQueue` / `SystemBuildingQueue` / `BuildQueue` mutation itself.
`communication/` should only schedule "when" a command arrives — the
"what happens on arrival" logic belongs next to the queue types.

New `macrocosmo/src/colony/remote.rs`:
- `pub fn apply_colony_command(...)` — the former private fn, unchanged
  behaviour, now with full doc comment on the arrival-time re-resolution
  contract.
- `pub type ApplyColoniesQuery` / `ApplySystemBuildingsQuery` — type
  aliases for the 4- and 2-tuple queries so the caller in
  `process_pending_commands` doesn't repeat them.
- `push_planet_building_order` remains a private helper.

`communication/mod.rs`:
- `process_pending_commands` now imports the type aliases and calls
  `crate::colony::apply_colony_command(...)`; its queue types stay
  exactly as before, same signature shape, no B0001 risk.
- Lost ~325 lines; gained 2 lines for the module pubs + helper calls.

No behaviour change. All workspace tests pass.

Closes #273.
